AW Customisation to ensure that the map is always positioned to the top, left and right edges of the app.
This is a workaround to changes in ionic to add safe space around the edge of the app, resulting in the map not being positioned correctly.
This also prevents the map from being hidden when forms slide in and out of the page as the delayed update between the native map and the webview no longer create a white boxed effect on the map.